separate app from server code
